criacao algumas entidade baseadas na documentacao parcial
