--- src/main/java/com/easy/chatbot/Logs/TransType.java
@@ -1,14 +1,0 @@
-package com.example.demo.Logs;
-
-public enum TransType {
-    PAY(-1), RECEIVE(1);
-
-    private final int type;
-    TransType(int type){
-        this.type = type;
-    }
-
-    public int getTypeMultiplier(){
-        return this.type;
-    }
-}
