+++ src/main/java/com/easy/chatbot/entitys/SQLDatabaseConnection.java
@@ -1,0 +1,28 @@
+package com.easy.chatbot.entitys;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class SQLDatabaseConnection {
+    // Connect to your database.
+    // Replace server name, username, and password with your credentials
+    public static void main(String[] args) {
+        String connectionUrl =
+                "jdbc:sqlserver://yourserver.database.windows.net:1433;"
+                        + "database=AdventureWorks;"
+                        + "user=yourusername@yourserver;"
+                        + "password=<password>;"
+                        + "encrypt=true;"
+                        + "trustServerCertificate=false;"
+                        + "loginTimeout=30;";
+
+        try (Connection connection = DriverManager.getConnection(connectionUrl);) {
+            // Code here.
+        }
+        // Handle any errors that may have occurred.
+        catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
+}
